pyproject.toml
[tool.poetry]
name = "oracle"
version = "0.1.0"
description = ""
authors = ["temcheeter <akravchyg10@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.15.0"
pillow = "^11.0.0"
pytz = "^2024.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
